- added new DI - refactored DIO - added support for Insecure HTTP connections
