+++ android/app/src/main/AndroidManifest.xml
@@ -14,6 +14,7 @@
          FlutterApplication and put your custom class here. -->
     <application
         android:icon="@mipmap/ic_launcher"
+        android:usesCleartextTraffic="true"
         android:label="Boilerplate Project">
         <activity
             android:name=".MainActivity"
