+++ lib/data/network/constants/endpoints.dart
@@ -3,13 +3,15 @@ class Endpoints {
 
   // base url
   static const String baseUrl = "http://jsonplaceholder.typicode.com";
+  // static const String baseUrl = "https://www.apphusetreach.no/application/119267";
 
   // receiveTimeout
-  static const int receiveTimeout = 5000;
+  static const int receiveTimeout = 15000;
 
   // connectTimeout
-  static const int connectionTimeout = 3000;
+  static const int connectionTimeout = 30000;
 
   // booking endpoints
   static const String getPosts = baseUrl + "/posts";
+  static const String getArticles = baseUrl + "/article/get_articles_list";
 }
