+++ lib/data/repository.dart
@@ -4,7 +4,6 @@ import 'package:boilerplate/data/local/datasources/post/post_datasource.dart';
 import 'package:boilerplate/data/sharedpref/shared_preference_helper.dart';
 import 'package:boilerplate/models/post/post.dart';
 import 'package:boilerplate/models/post/post_list.dart';
-import 'package:injectable/injectable.dart';
 import 'package:sembast/sembast.dart';
 
 import 'local/constants/db_constants.dart';
