+++ test/widget_test.dart
@@ -5,8 +5,7 @@
 // gestures. You can also use WidgetTester to find child widgets in the widget
 // tree, read text, and verify that the values of widget properties are correct.
 
-import 'package:boilerplate/main.dart';
-import 'package:flutter/material.dart';
+import 'package:boilerplate/ui/my_app.dart';
 import 'package:flutter_test/flutter_test.dart';
